Bdd yas

+++ src/main/java/com/briefBankApi/api/repository/ClientRepository.java
@@ -3,10 +3,11 @@ package com.briefBankApi.api.repository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import ch.qos.logback.core.net.server.Client;
+import com.briefBankApi.api.model.Client;
+
 
 @Repository
-public interface ClientRepository extends CrudRepository<Client, Integer> {
+public interface ClientRepository extends CrudRepository<Client, Long> {
 	
 
 }
