+++ src/main/java/com/briefBankApi/api/service/ClientService.java
@@ -6,11 +6,10 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.briefBankApi.api.model.Client;
 import com.briefBankApi.api.repository.ClientRepository;
 
-import ch.qos.logback.core.net.server.Client;
 import lombok.Data;
-
 
 @Data
 @Service
@@ -19,7 +18,7 @@ public class ClientService {
 	@Autowired
 	private ClientRepository clientRepository;
 	
-	public Optional<Client> getClientById (final Integer id) {
+	public Optional<Client> getClientById (final Long id) {
 		return clientRepository.findById(id);
 		
 	}
